refactor: change folder structure of components
